Move NormalizedState type + helpers to common-ui
