+++ src/common-ui/types.ts
@@ -1,0 +1,4 @@
+export interface NormalizedState<T> {
+    allIds: string[]
+    byId: { [id: string]: T }
+}
