+++ src/dashboard-refactor/search-results/util.ts
@@ -8,7 +8,6 @@ import {
 } from 'src/search/background/types'
 import {
     PageData,
-    NormalizedState,
     PageResult,
     PageResultsByDay,
     NoteData,
@@ -22,6 +21,7 @@ import {
 import { Annotation } from 'src/annotations/types'
 import { PAGE_SEARCH_DUMMY_DAY } from '../constants'
 import { sortByPagePosition } from 'src/sidebar/annotations-sidebar/sorting'
+import { initNormalizedState, mergeNormalizedStates } from 'src/common-ui/utils'
 
 export const notesTypeToString = (type: NotesType): string => {
     if (type === 'user') {
@@ -50,11 +50,6 @@ export const formatDayGroupTime = (day: number) =>
         lastWeek: '[Last] dddd',
         sameElse: 'dddd, DD MMMM, YYYY',
     })
-
-export const initNormalizedState = <T>(): NormalizedState<T> => ({
-    allIds: [],
-    byId: {},
-})
 
 export const getInitialFormState = (): NoteFormState => ({
     tags: [],
@@ -245,19 +240,6 @@ export const pageSearchResultToState: SearchResultToState = (
             },
         },
     }
-}
-
-export const mergeNormalizedStates = <T>(
-    ...toMerge: NormalizedState<T>[]
-): NormalizedState<T> => {
-    const merged: NormalizedState<T> = initNormalizedState()
-
-    for (const state of toMerge) {
-        merged.allIds = [...new Set([...merged.allIds, ...state.allIds])]
-        merged.byId = { ...merged.byId, ...state.byId }
-    }
-
-    return merged
 }
 
 export const mergeSearchResults = (
